web/src/utils/initStoryblok.ts
import { apiPlugin, storyblokInit } from "@storyblok/react";

import { publicRuntimeConfig } from "@/config";
import { logger } from "@/logger";

// Init connection to Storyblok
export const initStoryblok = (): void => {
	const accessToken = publicRuntimeConfig.storyblok.accessToken;
	const endpoint = publicRuntimeConfig.storyblok.ocelotEndpoint;
	const usingOcelotCache = !!endpoint;

	logger.info(
		{ endpoint, accessToken, usingOcelotCache },
		"initStoryblok test log order"
	);

	try {
		logger.info(
			`start initStoryblok with accessToken: ${accessToken} and endpoint: ${endpoint}`
		);

		logger.info(`initStoryblok usingOcelotCache: ${usingOcelotCache}`);

		storyblokInit({
			accessToken,
			use: [apiPlugin],
			apiOptions: {
				cache: {
					clear: usingOcelotCache ? "manual" : "auto",
					type: usingOcelotCache ? "none" : "memory",
				},
				endpoint,
			},
		});
	} catch (error) {
		logger.error(`Error initialising Storyblok: ${error}`, {
			ocelotEndpoint: endpoint,
			usingOcelotCache,
			originatingError: error instanceof Error && error.message,
			error,
		});
		new Error("Error initialising Storyblok", { cause: error });
	}
	logger.info("end initStoryblok");
};
